python/requirements.txt

# Common - REST interface
aiohttp==3.4.4
aiohttp_cors==0.7.0
cchardet==2.1.1
pyjwt==1.6.4

# Transformation of data, Apply JqFilter
pyjq==2.2.0

# openssl
pyopenssl==19.0.0
